Refactor code in preparation for parallelization
