+++ packages/mds-compliance/package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
     "start": "PATH_PREFIX=/compliance yarn watch server",
-    "historical": "yarn watch compute-historical-compliance",
+    "historical": "yarn watch historical/index.ts",
     "test": "yarn test:eslint && yarn test:unit",
     "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
     "test:unit": "PATH_PREFIX=/compliance DOTENV_CONFIG_PATH=../../.env nyc ts-mocha --project ../../tsconfig.json --exit",
